Trying to create a Creeper you can ride
It is in the game, but the game doesn't know what to do with it.
Also need to figure out how to replace vanilla creeper spawn with moddedcreeper spawn
Also need to figure out how to deactivate creeper attack mechanics with item, and allow it you to ride, wonder if I can have it extend creeperentity and passiveentity classes?